Cast a 32 bit durante el calculo de b

+++ auxiliares.c
@@ -30,11 +30,11 @@ void linear_map_init(int16_t *a, int32_t *b, enum boton bot){
 	switch (bot) {
 		case Boton_1: 
 			*a = (((int32_t)DCmaxBoton1<<5) - DCmin_S)/(BOTON1_MAX - BOTON1_MIN);
-			*b = DCmin_S - BOTON1_MIN * *a;
+			*b = DCmin_S - BOTON1_MIN * (int32_t)*a;
 			break;
 		case Boton_2:
 			*a = (((int32_t)DCmaxBoton2<<5) - DCmin_S)/(BOTON2_MAX - BOTON2_MIN);
-			*b = DCmin_S - BOTON2_MIN * *a;
+			*b = DCmin_S - BOTON2_MIN * (int32_t)*a;
 			break;
 	}
 }
